% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_year.R
\name{class_year}
\alias{class_year}
\title{Compute a student's high school graduation year from year and grade level}
\usage{
class_year(.end_year, .grade_level)
}
\arguments{
\item{.end_year}{\verb{<int>} the calendar year in the spring of a school year}

\item{.grade_level}{\verb{<fct>} a factor of grades in \verb{["PK", ..., "12"]}}
}
\value{
\verb{<int>} the year the kid should graduate from high school
}
\description{
Compute a student's high school graduation year from year and grade level
}
\examples{
class_year(2026L, factor(GRADE_LEVELS))

}
