% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_dates.R
\name{fix_dates}
\alias{fix_dates}
\title{Swap elements of date columns so that start dates always precede end dates}
\usage{
fix_dates(.enrollments, .start_field = "Start Date", .end_field = "End Date")
}
\arguments{
\item{.enrollments}{\verb{<tbl>} a table produced by \code{\link[=ingest_enrollments]{ingest_enrollments()}}}

\item{.start_field}{\verb{<chr?>} defaults to "Start Date"}

\item{.end_field}{\verb{<chr?>} defaults to "End Date"}
}
\value{
\verb{<tbl>} almost identical to \code{.enrollments} but with some start and end dates swapped
}
\description{
Swap elements of date columns so that start dates always precede end dates
}
