% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRADE_LEVELS}
\alias{GRADE_LEVELS}
\title{Grade Levels as a factor}
\format{
An object of class \code{character} of length 14.
}
\usage{
GRADE_LEVELS
}
\description{
fourteen two-character codes for grades, from PK to KG to 01 to 12
}
\keyword{datasets}
