% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hercrusdic-package.R
\docType{package}
\name{hercrusdic-package}
\alias{hercrusdic}
\alias{hercrusdic-package}
\title{hercrusdic: Use Infinite Campus Data from Racine Unified School District}
\description{
Since 2014-15, Racine Unified has used Infinite Campus as its student information system. This package provides an R interface to the kinds of csv files that one can download from Infinite Campus's web interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://higherx4racine.github.io/hercrusdic/}
}

}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@career2cradle.org}

Authors:
\itemize{
  \item Ben Taft \email{ben.taft@career2cradle.org}
}

Other contributors:
\itemize{
  \item Higher Expectations for Racine County [copyright holder]
}

}
\keyword{internal}
