% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_calendars.R
\name{ingest_calendars}
\alias{ingest_calendars}
\title{Calendars are unique combinations of a school and a school year}
\usage{
ingest_calendars(.db_dir)
}
\arguments{
\item{.db_dir}{\verb{<chr>} the full path to the database directory}
}
\value{
a tibble with 11 columns
\describe{
\item{Calendar ID}{\verb{<int>} the primary key for this table}
\item{End Year}{\verb{<int>} the year of the last date in this calendar}
\item{School ID}{\verb{<int>} a foreign key to the schools table}
\item{Start Date}{\verb{<date>} the date on which this calendar began}
\item{End Date}{\verb{<date>} the date on which this calendar finished}
\item{Student Minutes}{\verb{<int>} The duration of a full school day for this calendar}
\item{Whole Day Absence}{\verb{<int>} The threshold where missing some minutes becomes a full day absence}
\item{Half Day Absence}{\verb{<int>} The threshold where missing some minutes becomes a half day absence}
\item{Teacher Minutes}{\verb{<int>} The duration of a full teacher work day for this calendar}
\item{Is School Choice}{\verb{<lgl>} FALSE if this is a boundary, virtual, or non-voluntary school}
\item{Summer}{\verb{<lgl>} TRUE if the calendar represents a summer program, not full-year school}
}
}
\description{
Calendars are unique combinations of a school and a school year
}
