% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_enrollment_comments.R
\name{ingest_enrollment_comments}
\alias{ingest_enrollment_comments}
\title{Import enrollment records that have comments attached to them}
\usage{
ingest_enrollment_comments(.db_dir)
}
\arguments{
\item{.db_dir}{\verb{<chr>} the full path to the database directory}
}
\value{
a tibble with three columns
\describe{
\item{Enrollment ID}{\verb{<int>} a primary key shared with the enrollments table}
\item{Comments at Start}{\verb{<chr>} free text pertaining to the beginning of a student's enrollment}
\item{Comments at End}{\verb{<chr>} free text pertaining to the end of a student's enrollment}
}
}
\description{
Import enrollment records that have comments attached to them
}
