% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_enrollments.R
\name{ingest_enrollments}
\alias{ingest_enrollments}
\title{Import enrollments and sort them by student id and enrollment date}
\usage{
ingest_enrollments(.db_dir)
}
\arguments{
\item{.db_dir}{\verb{<chr>} the full path to the database folder}
}
\value{
a tibble with 9 columns
\describe{
\item{Enrollment ID}{\verb{<int>} the primary key}
\item{Student ID}{\verb{<int>} foreign key to the students table}
\item{Calendar ID}{\verb{<int>} foreign key to the calendars table}
\item{Grade Level}{\verb{<fct>} PK, KG, or 01 - 12}
\item{SPED}{\verb{<lgl>} TRUE if the enrollment is explicity noted as special education}
\item{Start Date}{\verb{<date>} when the enrollmen began}
\item{End Date}{\verb{<date>} when the enrollment ended}
\item{Bad Dates}{\verb{<lgl>} TRUE if \verb{Start Date < End Date} or \verb{is.na(End Date)}}
\item{Tier}{\verb{<fct>} Preschool, Elementary, Middle, or High, per \verb{Grade Level}}
\item{Order}{\verb{<int>} the position of an enrollment falls in a student's career}
}
}
\description{
Import enrollments and sort them by student id and enrollment date
}
