% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_grade_levels.R
\name{ingest_grade_levels}
\alias{ingest_grade_levels}
\title{Convert two-character grade codes into a factor of two-character grade levels}
\usage{
ingest_grade_levels(.grade_digraphs)
}
\arguments{
\item{.grade_digraphs}{\verb{<chr>} two-character codes with many different levels of pre-k}
}
\value{
a factor with all pre-k collapsed into "PK", then "KG" and "01"-"12"
}
\description{
Convert two-character grade codes into a factor of two-character grade levels
}
