% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_instructional_sections.R
\name{ingest_instructional_sections}
\alias{ingest_instructional_sections}
\title{Load a table of which sections correspond to, you know, actual classes}
\usage{
ingest_instructional_sections(.db_dir)
}
\arguments{
\item{.db_dir}{\verb{<chr>} the location of the database in the file system}
}
\value{
a tibble with 3 columns
\describe{
\item{Section ID}{\verb{<int>} the primary key for this table}
\item{Student ID}{\verb{<int>} a foreign key to the table of student demographic information}
\item{Calendar ID}{\verb{<int>} a foreign key to the many-to-many table of years and schools }
}
}
\description{
Load a table of which sections correspond to, you know, actual classes
}
