% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_schools.R
\name{ingest_schools}
\alias{ingest_schools}
\title{A school is a location or virtual administrative unit where students to learn}
\usage{
ingest_schools(.db_dir)
}
\arguments{
\item{.db_dir}{\verb{<chr>} the full path to the database directory}
}
\value{
a tibble with 8 columns
\describe{
\item{School ID}{\verb{<int>} the primary key for this table}
\item{NCES ID}{\verb{<chr>} six-digit id for this school in this state}
\item{State ID}{\verb{<int>} sex-digit id for this school in Wisconsin's db.}
\item{School Name}{\verb{<chr>} a human-readable, title-case label for each school}
\item{School Type}{\verb{<chr>} RUSD's internal code for the level of the school}
\item{Elementary}{\verb{<lgl>} \code{TRUE} if some of the school's students are in grades 4K-5}
\item{Middle}{\verb{<lgl>} \code{TRUE} if some of the school's students are in grades 6-8}
\item{High}{\verb{<lgl>} \code{TRUE} if some of the school's students are in grades 9-12}
}
}
\description{
A school is a location or virtual administrative unit where students to learn
}
