% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_students.R
\name{ingest_students}
\alias{ingest_students}
\title{Import demographic and life history data about students}
\usage{
ingest_students(.db_dir)
}
\arguments{
\item{.db_dir}{\verb{<chr>} the full path to the database directory}
}
\value{
a data frame with 13 columns
\describe{
\item{Student ID}{\verb{<int>} the primary key, internal to Infinite Camupus }
\item{Family Name}{\verb{<chr>} IN ALL CAPS}
\item{Given Name}{\verb{<chr>} ALSO IN ALL CAPS}
\item{Date of Birth}{\verb{<date>} useful for computing a kid's age…}
\item{Sex}{\verb{<chr>} "M" or "F"}
\item{Black}{\verb{<lgl>} TRUE if the student identifies as belonging to this race}
\item{Latine}{\verb{<lgl>}TRUE if the student identifies as belonging to this ethnicity}
\item{Student Number}{\verb{<int>} a six-digit unique used for many extramural reports}
\item{Native American}{\verb{<lgl>} TRUE if the student identifies as belonging to this race}
\item{Asian}{\verb{<lgl>} TRUE if the student identifies as belonging to this race}
\item{Pacific Islander}{\verb{<lgl>} TRUE if the student identifies as belonging to this race}
\item{Multiracial}{\verb{<lgl>} TRUE if the student identifies as belonging to this race}
\item{White}{\verb{<lgl>} TRUE if the student identifies as belonging to this race}
}
}
\description{
Import demographic and life history data about students
}
