% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delimited_table.R
\name{read_delimited_table}
\alias{read_delimited_table}
\title{Read caret-delimited text files}
\usage{
read_delimited_table(.path_to_table, .specification)
}
\arguments{
\item{.path_to_table}{\verb{<chr>} the full path to the table}

\item{.specification}{\verb{<lst>} a list of column types for \code{\link[readr:cols]{readr::cols()}}}
}
\value{
a tibble with features determined by \code{.specification}
}
\description{
Read caret-delimited text files
}
