% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TIERS_OF_EDUCATION}
\alias{TIERS_OF_EDUCATION}
\title{Map grade levels to tiers of education}
\format{
a tibble with 14 rows and two columns.
\describe{
\item{Grade Level}{\verb{<fct>} equal to \link{GRADE_LEVELS}}
\item{Tier}{\verb{<fct>} a factor with the four levels above}
}
}
\usage{
TIERS_OF_EDUCATION
}
\description{
\tabular{ll}{
   Grade Levels \tab Tier \cr
   PK \tab Preschool \cr
   KG-5 \tab Elementary \cr
   6-8 \tab Middle \cr
   9-12 \tab High \cr
}
}
\keyword{datasets}
